requirements.txt
fastybird-modules-metadata>=0.7
kink>=0.6
pony>=0.7
setuptools>=57.4
whistle>=1.0
